execution/_template.py
#!/usr/bin/env python3
"""
[Script Name]

Brief description of what this script does.

Usage:
    python execution/script_name.py --arg1 value1 --arg2 value2

Requirements:
    - Python 3.8+
    - Dependencies listed in requirements.txt

Environment Variables:
    - API_KEY: Description of required API key
    - OTHER_VAR: Description of other variable
"""

import os
import sys
import argparse
import logging
from pathlib import Path
from typing import Optional, Dict, Any
from dotenv import load_dotenv

# Configure logging
logging.basicConfig(
    level=logging.INFO,
    format='%(asctime)s - %(name)s - %(levelname)s - %(message)s'
)
logger = logging.getLogger(__name__)

# Load environment variables
load_dotenv()


class ScriptExecutor:
    """
    Main class for executing the script logic.

    Attributes:
        config: Configuration dictionary

    """

    def __init__(self, config: Dict[str, Any]):
        """
        Initialize the executor with configuration.

        Args:
            config: Dictionary containing configuration parameters
        """
        self.config = config
        self._validate_config()

    def _validate_config(self) -> None:
        """Validate that all required configuration is present."""
        required_keys = ['api_key']  # Add required keys here

        missing = [key for key in required_keys if key not in self.config]
        if missing:
            raise ValueError(f"Missing required configuration: {missing}")

    def execute(self) -> Dict[str, Any]:
        """
        Main execution method.

        Returns:
            Dictionary containing execution results

        Raises:
            Exception: If execution fails
        """
        try:
            logger.info("Starting execution...")

            # Add your main logic here
            result = self._process()

            logger.info("Execution completed successfully")
            return result

        except Exception as e:
            logger.error(f"Execution failed: {str(e)}")
            raise

    def _process(self) -> Dict[str, Any]:
        """
        Core processing logic.

        Returns:
            Processing results
        """
        # Implement your core logic here
        return {
            'status': 'success',
            'data': {}
        }


def parse_arguments() -> argparse.Namespace:
    """
    Parse command-line arguments.

    Returns:
        Parsed arguments namespace
    """
    parser = argparse.ArgumentParser(
        description='Description of what this script does'
    )

    parser.add_argument(
        '--arg1',
        type=str,
        required=True,
        help='Description of argument 1'
    )

    parser.add_argument(
        '--arg2',
        type=str,
        default='default_value',
        help='Description of argument 2'
    )

    parser.add_argument(
        '--output',
        type=str,
        default='.tmp/output.json',
        help='Output file path'
    )

    parser.add_argument(
        '--verbose',
        action='store_true',
        help='Enable verbose logging'
    )

    return parser.parse_args()


def main() -> int:
    """
    Main entry point for the script.

    Returns:
        Exit code (0 for success, 1 for failure)
    """
    try:
        # Parse arguments
        args = parse_arguments()

        # Set logging level
        if args.verbose:
            logging.getLogger().setLevel(logging.DEBUG)

        # Prepare configuration
        config = {
            'api_key': os.getenv('API_KEY'),
            'arg1': args.arg1,
            'arg2': args.arg2,
            'output_path': args.output
        }

        # Execute
        executor = ScriptExecutor(config)
        result = executor.execute()

        # Save output if needed
        output_path = Path(args.output)
        output_path.parent.mkdir(parents=True, exist_ok=True)

        logger.info(f"Results: {result}")
        logger.info(f"Output saved to: {output_path}")

        return 0

    except Exception as e:
        logger.error(f"Script failed: {str(e)}")
        return 1


if __name__ == '__main__':
    sys.exit(main())
