requirements.txt
# Core dependencies
python-dotenv>=1.0.0

# HTTP and API clients
requests>=2.31.0

# Google API clients (if needed)
# google-api-python-client>=2.0.0
# google-auth-httplib2>=0.1.0
# google-auth-oauthlib>=1.0.0

# Data processing
# pandas>=2.0.0
# numpy>=1.24.0

# Web scraping (if needed)
# beautifulsoup4>=4.12.0
# selenium>=4.0.0

# Database (if needed)
# psycopg2-binary>=2.9.0  # PostgreSQL
# pymongo>=4.0.0  # MongoDB
# redis>=4.5.0  # Redis

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Code quality
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0

# Utilities
pyyaml>=6.0.0
click>=8.1.0
